+++ Plugin/src/CodeRebirthConfig.cs
@@ -7,6 +7,7 @@ namespace CodeRebirth.src;
 public class CodeRebirthConfig
 {
     #region Enables/Disables
+    public ConfigEntry<bool> ConfigFirstLaunchPopup { get; private set; }
     public ConfigEntry<bool> ConfigFunctionalMicrowaveEnabled { get; private set; }
     public ConfigEntry<bool> ConfigInsideBearTrapEnabled { get; private set; }
     public ConfigEntry<bool> ConfigBearTrapEnabled { get; private set; }
@@ -624,6 +625,10 @@ public class CodeRebirthConfig
                                             "Min and Max value of the Emerald, leave at -1 for both defaults to not mess with base values, values are NOT multiplied by 0.4.");
         #endregion
         #region ModCompat
+        ConfigFirstLaunchPopup = configFile.Bind("ModCompat Options",
+                                            "First Launch Popup",
+                                            true,
+                                            "Enables/Disables the first launch popup for the host about the ShockwaveGalModelReplacement mod.");
         #endregion
         #region CutieFly
         ConfigCutieFlyEnabled = configFile.Bind("CutieFly Options",