+++ Plugin/src/Patches/PlayerControllerBPatch.cs
@@ -33,12 +33,25 @@ static class PlayerControllerBPatch {
         __instance.AddCRPlayerData();
     }
 
-    public static void Init() {
+    public static void Init()
+    {
+        On.GameNetcodeStuff.PlayerControllerB.ConnectClientToPlayerObject += PlayerControllerB_ConnectClientToPlayerObject;
         On.GameNetcodeStuff.PlayerControllerB.TeleportPlayer += PlayerControllerB_TeleportPlayer;
         On.GameNetcodeStuff.PlayerControllerB.DamagePlayer += PlayerControllerB_DamagePlayer;
         IL.GameNetcodeStuff.PlayerControllerB.CheckConditionsForSinkingInQuicksand += PlayerControllerB_CheckConditionsForSinkingInQuicksand;
         // IL.GameNetcodeStuff.PlayerControllerB.DiscardHeldObject += ILHookAllowParentingOnEnemy_PlayerControllerB_DiscardHeldObject;
         On.GameNetcodeStuff.PlayerControllerB.LateUpdate += PlayerControllerB_LateUpdate;
+    }
+
+    private static void PlayerControllerB_ConnectClientToPlayerObject(On.GameNetcodeStuff.PlayerControllerB.orig_ConnectClientToPlayerObject orig, PlayerControllerB self)
+    {
+        orig(self);
+        Plugin.ExtendedLogging("PlayerControllerB_ConnectClientToPlayerObject called");
+        if (self.IsServer && Plugin.ModConfig.ConfigFirstLaunchPopup.Value && (!Plugin.ModelReplacementAPIIsOn || !Plugin.MoreSuitsIsOn))
+        {
+            HUDManager.Instance.DisplayTip("Mod not detected", "Downloading ModelReplacementAPI and MoreSuits adds a new suit as the ShockwaveGal's model");
+            Plugin.ModConfig.ConfigFirstLaunchPopup.Value = false;
+        }
     }
 
     private static void PlayerControllerB_DamagePlayer(On.GameNetcodeStuff.PlayerControllerB.orig_DamagePlayer orig, PlayerControllerB self, int damageNumber, bool hasDamageSFX, bool callRPC, CauseOfDeath causeOfDeath, int deathAnimation, bool fallDamage, Vector3 force)